+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## Version 0.7
 xx.xx.xxxx
 
-- Export to Eclipse Vorto
+- Export to Eclipse Vorto, first version
 - First version of export to SVG in oneM2M resource format
 - Fixed errors in Java export
 - Fixed errors in OPML export
 - Improved markdown export
+- Added optional reading of command line arguments from configuration file(s)
 - Made SDT3 the default input format
 - minor bug fixes